Update WeatherViewModel.java
Uses a zip code from Seattle. Note that the API key used for the time being comes from this site: https://www.zipcodeapi.com/API#zipToLoc
The API key on this site seems to update on a daily basis. So if you plan on continuing to use this, please be sure to check on the webpage and get the updated key. This also seems to have a 10 call/hour restriction.